refactor: 阶段2 - 脚本模块化 + 工程化
- video_recap.py (3297行) 拆分为 11 个模块文件:
  config.py / common.py / extract.py / detect.py / asr.py
  vlm.py / narration.py / tts.py / assemble.py / pipeline.py
  video_recap.py 瘦身到 96 行 (main + argparse)
- CLI 入口/参数/输出/resume 标记零变化
- 新增 Makefile (recap/demo/lint/test/clean)
- 新增 GitHub Actions (lint + SKILL.md + prompt锚点校验)
- 新增 tests/test_pure.py (纯函数单测)
- internal-config.md 顶部已加 CONFIG 反向引用
