+++ skills/video-recap/scripts/common.py
@@ -1,0 +1,137 @@
+import json
+import re
+import subprocess
+import time
+import urllib.request
+import urllib.error
+
+from config import CONFIG, PROMPTS_DIR
+
+# ── 工具函数 ──────────────────────────────────────────────────────────
+
+def log(msg):
+    print(f"[video-recap] {msg}", flush=True)
+
+
+def run_cmd(cmd, **kwargs):
+    """运行命令，返回 CompletedProcess"""
+    log(f"运行: {' '.join(str(c) for c in cmd) if isinstance(cmd, list) else cmd}")
+    return subprocess.run(cmd, capture_output=True, text=True, **kwargs)
+
+
+def get_video_duration(video_path):
+    """获取视频时长（秒）"""
+    cmd = ["ffprobe", "-v", "quiet", "-show_entries", "format=duration",
+           "-of", "csv=p=0", str(video_path)]
+    result = run_cmd(cmd)
+    if result.returncode != 0:
+        return 0.0
+    return float(result.stdout.strip())
+
+
+def api_call(payload, max_retries=5):
+    """调用 OpenAI-compatible API，带重试"""
+    headers = {
+        "Authorization": f"Bearer {CONFIG['api_key']}",
+        "Content-Type": "application/json",
+        "User-Agent": "video-recap/1.0",
+    }
+    data = json.dumps(payload).encode("utf-8")
+
+    for attempt in range(max_retries):
+        try:
+            req = urllib.request.Request(CONFIG["api_url"], data=data, headers=headers)
+            with urllib.request.urlopen(req, timeout=300) as resp:
+                result = json.loads(resp.read().decode("utf-8"))
+                return result
+        except urllib.error.HTTPError as e:
+            body = e.read().decode("utf-8", errors="replace")[:500]
+            wait = 2 ** attempt
+            if e.code == 429:
+                retry_after = e.headers.get("Retry-After")
+                if retry_after:
+                    wait = max(wait, int(retry_after))
+                log(f"API 速率限制 (尝试 {attempt+1}/{max_retries}), 等待 {wait}s")
+            elif e.code == 401:
+                raise RuntimeError("API 认证失败 (401)。请检查 OPENAI_API_KEY 是否正确。")
+            elif e.code == 403:
+                hint = "API 访问被拒绝 (403)。"
+                if "1010" in body or "cloudflare" in body.lower():
+                    hint += "IP 被 Cloudflare 限流，请等待几分钟后重试。"
+                    raise RuntimeError(hint)
+                hint += "请检查 API key 权限和 OPENAI_API_URL 设置。"
+                raise RuntimeError(hint)
+            elif e.code == 405:
+                raise RuntimeError(f"API 端点不可用 (405)，可能被 WAF 拦截。请检查 OPENAI_API_URL 或稍后重试。")
+            elif e.code == 503:
+                log(f"API 服务暂不可用 (503)，等待 {wait}s (尝试 {attempt+1}/{max_retries})")
+            elif e.code == 524:
+                # Cloudflare 超时：服务端处理超时，需要更长退避
+                wait = max(wait, 4 * (attempt + 1))
+                log(f"API 超时 (524)，等待 {wait}s (尝试 {attempt+1}/{max_retries})")
+            else:
+                log(f"API 调用失败 (尝试 {attempt+1}/{max_retries}): HTTP {e.code} — {body}")
+            if attempt < max_retries - 1:
+                time.sleep(wait)
+            else:
+                raise RuntimeError(f"API 调用失败 {max_retries} 次: HTTP {e.code} — {body}")
+        except (urllib.error.URLError, Exception) as e:
+            wait = 2 ** attempt
+            log(f"API 调用失败 (尝试 {attempt+1}/{max_retries}): {e}")
+            if attempt < max_retries - 1:
+                log(f"等待 {wait}s 后重试...")
+                time.sleep(wait)
+            else:
+                raise RuntimeError(f"API 调用失败 {max_retries} 次: {e}")
+
+
+def load_prompt(name):
+    """加载 prompt 模板"""
+    path = PROMPTS_DIR / "prompt-templates.md"
+    if not path.exists():
+        return None
+    content = path.read_text()
+    # 用 ### NAME 和 ### 分隔提取对应 prompt
+    pattern = rf"### {name}\s*\n(.*?)(?=\n### |\Z)"
+    m = re.search(pattern, content, re.DOTALL)
+    return m.group(1).strip() if m else None
+
+
+
+def _parse_narration_json(text):
+    """从 LLM 输出中解析 JSON 解说脚本"""
+    if not text or not text.strip():
+        return []
+
+    # 尝试直接解析
+    try:
+        result = json.loads(text)
+        if isinstance(result, list):
+            return result
+        if isinstance(result, dict):
+            return [result]
+    except json.JSONDecodeError:
+        pass
+
+    # 尝试提取 JSON 块（数组或对象）
+    patterns = [
+        r"```json\s*(.*?)\s*```",
+        r"```\s*(.*?)\s*```",
+        r"(\[.*\])",
+        r"(\{.*\})",
+    ]
+    for pattern in patterns:
+        m = re.search(pattern, text, re.DOTALL)
+        if m:
+            try:
+                result = json.loads(m.group(1))
+                if isinstance(result, list) and len(result) > 0:
+                    return result
+                if isinstance(result, dict):
+                    return [result]
+            except json.JSONDecodeError:
+                continue
+
+    log(f"警告: 无法解析 LLM JSON 输出 ({len(text)} 字符)")
+    return []
+
