+++ skills/video-recap/tests/test_pure.py
@@ -1,0 +1,30 @@
+"""Pure function unit tests for video-recap modules."""
+import sys
+import os
+
+sys.path.insert(0, os.path.join(os.path.dirname(__file__), '..', 'scripts'))
+
+from narration import _text_char_count
+from tts import _parse_rate_offset
+from assemble import _seconds_to_srt_time
+
+
+def test_text_char_count():
+    assert _text_char_count("hello") == 5
+    assert _text_char_count("你好世界") == 4
+    assert _text_char_count("") == 0
+
+
+def test_parse_rate_offset():
+    assert _parse_rate_offset("+0%") == 0.0
+    assert _parse_rate_offset("+20%") == 0.2
+    assert _parse_rate_offset("-10%") == -0.1
+    assert _parse_rate_offset("+5%") == 0.05
+
+
+def test_seconds_to_srt_time():
+    # 3661.5s = 1h 1m 1.5s
+    result = _seconds_to_srt_time(3661.5)
+    assert result.startswith("01:01:01")
+    # 0s
+    assert _seconds_to_srt_time(0) == "00:00:00,000"
